hwtracing: owl: Use 16 bits for metadata cpu field

+++ drivers/hwtracing/owl/tracectrl.c
@@ -490,7 +490,7 @@ static void tracectrl_insert_metadata(struct tracectrl *ctrl,
 
 	entry = &ctrl->metadata[ctrl->used_metadata_entries];
 	entry->timestamp = get_timestamp();
-	entry->cpu = (u8) smp_processor_id();
+	entry->cpu = (u16) smp_processor_id();
 	entry->pid = (int) task_pid_nr(task);
 	entry->ppid = (int) task_ppid_nr(task);
 	entry->has_mm = task->mm != NULL;
